requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
astropy==3.0.5
atomicwrites==1.3.0
attrs==19.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.2.0
boto==2.49.0
boto3==1.9.161
botocore==1.12.161
Bottleneck==1.2.1
bz2file==0.98
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
colorama==0.4.1
comtypes==1.1.7
dask==1.2.0
decorator==4.4.0
defusedxml==0.5.0
distributed==1.28.1
dj-database-url==0.5.0
Django==2.2.1
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.10
Flask==1.0.3
future==0.17.1
gast==0.2.2
gensim==3.7.3
gevent==1.4.0
glob2==0.6
greenlet==0.4.15
grpcio==1.20.1
gunicorn==19.9.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jmespath==0.9.4
joblib==0.13.2
libarchive-c==2.8
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mock==3.0.5
more-itertools==4.3.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.2.5
numba==0.43.1
oauthlib==3.0.1
olefile==0.46
openpyxl==2.6.1
packaging==19.0
pandas==0.23.4
parso==0.4.0
partd==0.3.9
path.py==12.0.1
pathlib2==2.3.3
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
ply==3.11
protobuf==3.7.1
psutil==5.6.2
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.15.2
PySocks==1.7.0
pytest==4.6.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-crfsuite==0.9.6
python-dateutil==2.8.0
pytz==2019.1
rope==0.14.0
scikit-image==0.14.2
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.1
spyder==3.3.4
sympy==1.4
tables==3.4.4
tblib==1.4.0
termcolor==1.1.0
terminado==0.8.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.32.1
typed-ast==1.3.5
unicodecsv==0.14.1
urllib3==1.24.3
wcwidth==0.1.7
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
zict==0.1.4
zipp==0.5.1